requirements.txt
py-cord[speed]
PyYAML
RDFLib
